Create shim implementations of libNiFpgaLv.so.13 and libnirio_emb_can.so.16
Netcomm depends on emb_can and ChipObject depends on FpgaLv.

emb_can pulls in a few dependencies, and is only used for 6 functions.

NiFpgaLv is only used for a single function that is never actually used, however it pulls in a TON of dependencies. Some of these dependencies have missing implementations. We provide shim implementations of these functions, however in all cases, these shims are broken, and if they were actually called, things would break. They are never called, unless tSystemInterface::getLVHandle() is called, but even today with our shims that function is horrible broken.

Replace both of these libraries with shims, that only include the minimal functions needed. These files are never deployed to the roborio, so are only used for linking purposes